Add mine strategy

+++ src/server/strategies/movestrategy.cpp
@@ -1,7 +1,8 @@
-#include "../performers/movingperformer.h"
 #include "movestrategy.h"
 
-void server::MoveStrategy::tick(std::shared_ptr<core::GameWorld> world, double timeDelta) {
+#include "../performers/movingperformer.h"
+
+void server::MoveStrategy::tick(std::shared_ptr<core::GameWorld> world, int timeDelta) {
     moving_performer::moveIfNoObstacles(getObject(), timeDelta, world, moving);
 }
 
@@ -9,7 +10,7 @@ QString server::MoveStrategy::getName() {
     return name;
 }
 
-server::MoveStrategy::MoveStrategy(std::shared_ptr<core::Object> object) : Strategy(object) {}
+server::MoveStrategy::MoveStrategy(std::shared_ptr<core::Object> object): Strategy(object) {}
 
 void server::MoveStrategy::assign(server::DataBundle& dataBundle) {
     dataBundle.assign("moving", moving);
